PRC-443: Apply SecureAPIIntegrationTestBase to all API tests
Except the deprecated reference data APIs.
