+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/integration/resource/UpdatePrisonerContactRestrictionIntegrationTest.kt
@@ -9,8 +9,9 @@ import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.ValueSource
 import org.springframework.http.MediaType
+import org.springframework.test.web.reactive.server.WebTestClient
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.client.manage.users.User
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.integration.PostgresIntegrationTestBase
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.integration.SecureAPIIntegrationTestBase
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactRelationship
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreatePrisonerContactRestrictionRequest
@@ -22,11 +23,13 @@ import uk.gov.justice.digital.hmpps.hmppscontactsapi.service.events.Source
 import uk.gov.justice.hmpps.kotlin.common.ErrorResponse
 import java.time.LocalDate
 
-class UpdatePrisonerContactRestrictionIntegrationTest : PostgresIntegrationTestBase() {
+class UpdatePrisonerContactRestrictionIntegrationTest : SecureAPIIntegrationTestBase() {
   private var savedContactId = 0L
   private var savedPrisonerContactId = 0L
   private var savedPrisonerContactRestrictionId = 0L
   private val prisonerNumberCreatedAgainst = "A1234AA"
+
+  override val allowedRoles: Set<String> = setOf("ROLE_CONTACTS_ADMIN", "ROLE_CONTACTS__RW")
 
   @BeforeEach
   fun initialiseData() {
@@ -62,43 +65,11 @@ class UpdatePrisonerContactRestrictionIntegrationTest : PostgresIntegrationTestB
     ).prisonerContactRestrictionId
   }
 
-  @Test
-  fun `should return unauthorized if no token`() {
-    webTestClient.put()
-      .uri("/prisoner-contact/$savedPrisonerContactId/restriction/$savedPrisonerContactRestrictionId")
-      .accept(MediaType.APPLICATION_JSON)
-      .contentType(MediaType.APPLICATION_JSON)
-      .bodyValue(aMinimalRequest())
-      .exchange()
-      .expectStatus()
-      .isUnauthorized
-  }
-
-  @Test
-  fun `should return forbidden if no role`() {
-    webTestClient.put()
-      .uri("/prisoner-contact/$savedPrisonerContactId/restriction/$savedPrisonerContactRestrictionId")
-      .accept(MediaType.APPLICATION_JSON)
-      .contentType(MediaType.APPLICATION_JSON)
-      .bodyValue(aMinimalRequest())
-      .headers(setAuthorisation())
-      .exchange()
-      .expectStatus()
-      .isForbidden
-  }
-
-  @Test
-  fun `should return forbidden if wrong role`() {
-    webTestClient.put()
-      .uri("/prisoner-contact/$savedPrisonerContactId/restriction/$savedPrisonerContactRestrictionId")
-      .accept(MediaType.APPLICATION_JSON)
-      .contentType(MediaType.APPLICATION_JSON)
-      .bodyValue(aMinimalRequest())
-      .headers(setAuthorisation(roles = listOf("ROLE_WRONG")))
-      .exchange()
-      .expectStatus()
-      .isForbidden
-  }
+  override fun baseRequestBuilder(): WebTestClient.RequestHeadersSpec<*> = webTestClient.put()
+    .uri("/prisoner-contact/$savedPrisonerContactId/restriction/$savedPrisonerContactRestrictionId")
+    .accept(MediaType.APPLICATION_JSON)
+    .contentType(MediaType.APPLICATION_JSON)
+    .bodyValue(aMinimalRequest())
 
   @ParameterizedTest
   @CsvSource(
